Regroup pathutil functions.
