+++ pathutil/path_filter.go
@@ -9,6 +9,27 @@ import (
 	"strings"
 )
 
+// ListEntries filters contents of a directory using the provided filters
+func ListEntries(dir string, filters ...FilterFunc) ([]string, error) {
+	absDir, err := filepath.Abs(dir)
+	if err != nil {
+		return []string{}, err
+	}
+
+	entries, err := ioutil.ReadDir(absDir)
+	if err != nil {
+		return []string{}, err
+	}
+
+	var paths []string
+	for _, entry := range entries {
+		pth := filepath.Join(absDir, entry.Name())
+		paths = append(paths, pth)
+	}
+
+	return FilterPaths(paths, filters...)
+}
+
 // FilterPaths ...
 func FilterPaths(fileList []string, filters ...FilterFunc) ([]string, error) {
 	var filtered []string
