+++ pathutil/sortable_path.go
@@ -7,6 +7,37 @@ import (
 	"strings"
 )
 
+// ListPathInDirSortedByComponents ...
+func ListPathInDirSortedByComponents(searchDir string, relPath bool) ([]string, error) {
+	searchDir, err := filepath.Abs(searchDir)
+	if err != nil {
+		return []string{}, err
+	}
+
+	var fileList []string
+
+	if err := filepath.Walk(searchDir, func(path string, _ os.FileInfo, walkErr error) error {
+		if walkErr != nil {
+			return walkErr
+		}
+
+		if relPath {
+			rel, err := filepath.Rel(searchDir, path)
+			if err != nil {
+				return err
+			}
+			path = rel
+		}
+
+		fileList = append(fileList, path)
+
+		return nil
+	}); err != nil {
+		return []string{}, err
+	}
+	return SortPathsByComponents(fileList)
+}
+
 // SortablePath ...
 type SortablePath struct {
 	Pth        string
